Use ColumnName component in FilterEntry and GroupEntry components
