+++ mathesar_ui/src/typeUtils.ts
@@ -1,1 +1,8 @@
 export type Mutable<T> = { -readonly [P in keyof T]: T[P] };
+
+export interface ReadableMapLike<Key, Value> {
+  size: number;
+  get: (key: Key) => Value | undefined;
+  keys(): IterableIterator<Key>;
+  values(): IterableIterator<Value>;
+}
